Fixed right clicking on the "Run" link
The "href" attribute was previously changed to the data URL only when "click" or "auxclick" fired, but a "mousedown" is definitely an indication that the user is going to click on it. By making the "mousedown" change the "href" attribute, right clicking is available to the user for opening in a new tab or window.